fix primaryAsset vs targetAsset issue
