+++ src/components/Compounder/index.ts
@@ -1,6 +1,5 @@
 export * from "./CompounderVaultApy"
 export * from "./CompounderVaultDepositForm"
-export * from "./CompounderVaultName"
 export * from "./CompounderVaultTable"
 export * from "./CompounderVaultTvl"
 export * from "./CompounderVaultUserBalance"
