+++ src/components/Concentrator/ConcentratorVaultWithdrawForm.tsx
@@ -8,20 +8,20 @@ import { VaultWithdrawForm } from "@/components/VaultRow/lib"
 
 export const ConcentratorVaultWithdrawForm: FC<VaultProps> = (props) => {
   const concentrator = useConcentratorVault({
-    targetAsset: props.vaultAddress,
-    primaryAsset: props.asset,
+    targetAsset: props.asset,
+    primaryAsset: props.vaultAddress,
     type: props.type ?? "balancer",
   })
 
   const { data: underlyingAssets } = useConcentratorUnderlyingAssets({
-    targetAsset: props.vaultAddress,
-    primaryAsset: props.asset,
+    targetAsset: props.asset,
+    primaryAsset: props.vaultAddress,
   })
   return (
     <VaultWithdrawForm
       {...props}
       defaultInputToken={concentrator?.data?.ybTokenAddress}
-      defaultOutputToken={props.asset}
+      defaultOutputToken={props.vaultAddress}
       underlyingAssets={underlyingAssets}
       productType="concentrator"
     />
