+++ src/components/Concentrator/index.ts
@@ -3,7 +3,6 @@ export * from "./ConcentratorRewards"
 export * from "./ConcentratorTargetAsset"
 export * from "./ConcentratorVaultApy"
 export * from "./ConcentratorVaultDepositForm"
-export * from "./ConcentratorVaultName"
 export * from "./ConcentratorVaultTable"
 export * from "./ConcentratorVaultTvl"
 export * from "./ConcentratorVaultUserBalance"
