+++ src/hooks/useConcentratorVaultTvl.ts
@@ -2,13 +2,13 @@ import { VaultProps } from "@/lib/types"
 import useConcentratorVaultTvlFallback from "@/hooks/lib/tvl/concentrator/useConcentratorVaultTvlFallback"
 
 export function useConcentratorVaultTvl({
-  asset: primaryAsset,
-  vaultAddress: targetAsset,
+  asset: targetAsset,
+  vaultAddress: primaryAsset,
   type,
 }: VaultProps) {
   return useConcentratorVaultTvlFallback({
-    asset: primaryAsset,
-    vaultAddress: targetAsset,
+    targetAsset,
+    primaryAsset,
     type,
   })
 }
